+++ client/__init__.py
@@ -1,5 +1,0 @@
-from client.client import APIClient
-
-__version__ = "1.0.0"
-
-__all__ = ["APIClient"]