+++ client/utils/__init__.py
@@ -1,4 +1,0 @@
-from client.utils.base_config import BaseConfig
-from client.utils.schemas import APIClientSettings
-
-__all__ = ["BaseConfig", "APIClientSettings"]