+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## Changelog
 
+### v1.0.1
+_13/02/2025_
+
+- Fix module name from `client` to `bcn_rainfall_api_client`
+- Allow to pass `path` parameter directly in `APIClient.from_config` class method
+
 ### v1.0.0 
 _13/02/2025_